+++ tests/ClassesAndObjects/HasObjectPropertiesTest.php
@@ -1,0 +1,160 @@
+<?php
+
+/**
+ * Copyright (c) 2016-present Ganbaro Digital Ltd
+ * All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions
+ * are met:
+ *
+ *   * Redistributions of source code must retain the above copyright
+ *     notice, this list of conditions and the following disclaimer.
+ *
+ *   * Redistributions in binary form must reproduce the above copyright
+ *     notice, this list of conditions and the following disclaimer in
+ *     the documentation and/or other materials provided with the
+ *     distribution.
+ *
+ *   * Neither the names of the copyright holders nor the names of his
+ *     contributors may be used to endorse or promote products derived
+ *     from this software without specific prior written permission.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+ * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+ * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
+ * FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE
+ * COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT,
+ * INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
+ * BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
+ * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
+ * CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT
+ * LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
+ * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
+ * POSSIBILITY OF SUCH DAMAGE.
+ *
+ * @category  Libraries
+ * @package   MissingBits/ClassesAndObjects
+ * @author    Stuart Herbert <stuherbert@ganbarodigital.com>
+ * @copyright 2016-present Ganbaro Digital Ltd www.ganbarodigital.com
+ * @license   http://www.opensource.org/licenses/bsd-license.php  BSD License
+ * @link      http://ganbarodigital.github.io/php-the-missing-bits
+ */
+
+ namespace GanbaroDigitalTest\MissingBits\ClassesAndObjects;
+
+ use GanbaroDigital\MissingBits\ClassesAndObjects\HasObjectProperties;
+ use InvalidArgumentException;
+ use PHPUnit_Framework_TestCase;
+ use stdClass;
+
+ /**
+  * @coversDefaultClass GanbaroDigital\MissingBits\ClassesAndObjects\HasObjectProperties
+  */
+class HasObjectPropertiesTest extends PHPUnit_Framework_TestCase
+{
+    /**
+     * @covers ::check
+     */
+    public function test_returns_false_if_object_has_no_non_static_properties()
+    {
+        // ----------------------------------------------------------------
+        // setup your test
+
+        // ----------------------------------------------------------------
+        // perform the change
+
+        $actualResult = HasObjectProperties::check(new HasObjectPropertiesTest_NoNonStaticProperties);
+
+        // ----------------------------------------------------------------
+        // test the results
+
+        $this->assertFalse($actualResult);
+    }
+
+    /**
+     * @covers ::check
+     */
+    public function test_returns_true_if_object_has_non_static_properties()
+    {
+        // ----------------------------------------------------------------
+        // setup your test
+
+        // ----------------------------------------------------------------
+        // perform the change
+
+        $actualResult = HasObjectProperties::check(new HasObjectPropertiesTest_SeveralNonStaticProperties);
+
+        // ----------------------------------------------------------------
+        // test the results
+
+        $this->assertTrue($actualResult);
+    }
+
+    /**
+     * @covers ::check
+     * @dataProvider provideNonObjects
+     */
+    public function test_throws_InvalidArgumentException_for_non_objects($target, $expectedType)
+    {
+        // ----------------------------------------------------------------
+        // setup your test
+
+        $expectedMessage = '$target is not an object, is a ' . $expectedType;
+        $actualMessage = null;
+
+        // ----------------------------------------------------------------
+        // perform the change
+
+        try {
+            HasObjectProperties::check($target);
+        }
+        catch (InvalidArgumentException $e) {
+            $actualMessage = $e->getMessage();
+        }
+
+        // ----------------------------------------------------------------
+        // test the results
+
+        // this will only pass:
+        //
+        // 1 - if an exception was thrown at all, and
+        // 2 - if the exception contained the correct message
+        //
+        // this helps us catch cases where the right exception is being
+        // thrown for other reasons
+        $this->assertEquals($expectedMessage, $actualMessage);
+    }
+
+    public function provideNonObjects()
+    {
+        return [
+            [ null, 'NULL' ],
+            [ [], 'array' ],
+            [ true, 'boolean<true>' ],
+            [ false, 'boolean<false>' ],
+            [ 0.0, 'double<0>' ],
+            [ -100.0, 'double<-100>' ],
+            [ 100.0, 'double<100>' ],
+            [ 3.1415927, 'double<3.1415927>' ],
+            [ 0, 'integer<0>' ],
+            [ -100, 'integer<-100>' ],
+            [ 100, 'integer<100>' ],
+            [ STDIN, 'resource' ],
+            [ 'doesNotExist', 'string<doesNotExist>' ],
+        ];
+    }
+}
+
+class HasObjectPropertiesTest_NoNonStaticProperties
+{
+    public static $staticProp1 = 1;
+}
+
+class HasObjectPropertiesTest_SeveralNonStaticProperties
+{
+    public $objProp1 = 1;
+    public $objProp2 = 2;
+    public $objProp3 = 3;
+    public $objProp4;
+}